feat: fix import path issues and update FSD structure
Consistently update all import paths to use `~/shared/ui/*`
Update Next.js and TypeScript configurations for path aliases
Ensure all UI components load correctly with FSD import paths

#VERCEL_SKIP
