+++ package.json
@@ -1,14 +1,12 @@
 {
   "name": "irving-portfolio-fsd",
   "version": "0.1.0",
-  "type": "module",
+  "private": true,
   "scripts": {
-    "astro": "astro",
-    "build": "astro check && astro build",
-    "check": "astro check",
-    "dev": "astro dev",
-    "preview": "astro preview",
-    "start": "astro dev"
+    "build": "next build",
+    "dev": "next dev",
+    "lint": "next lint",
+    "start": "next start"
   },
   "dependencies": {
     "@astrojs/react": "latest",
@@ -23,19 +21,22 @@
     "effector": "latest",
     "effector-react": "latest",
     "lucide-react": "^0.454.0",
-    "next": "15.2.4",
+    "next": "14.2.16",
     "next-themes": "latest",
-    "react": "^19",
-    "react-dom": "^19",
+    "react": "^18",
+    "react-dom": "^18",
     "tailwind-merge": "^2.5.5",
     "tailwindcss": "^3.4.17",
-    "tailwindcss-animate": "^1.0.7"
+    "tailwindcss-animate": "^1.0.7",
+    "typescript": "^5.0.0"
   },
   "devDependencies": {
-    "@astrojs/check": "^0.3.0",
     "@types/node": "^22",
-    "@types/react": "^19",
-    "@types/react-dom": "^19",
+    "@types/react": "^18",
+    "@types/react-dom": "^18",
+    "autoprefixer": "^10.0.0",
+    "eslint": "^8.0.0",
+    "eslint-config-next": "14.0.0",
     "postcss": "^8.5",
     "typescript": "^5"
   }
