+++ widgets/skills-section.tsx
@@ -1,6 +1,6 @@
-import { Badge } from "@/components/ui/badge"
-import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
-import { Progress } from "@/components/ui/progress"
+import { Badge } from "~/shared/ui/badge"
+import { Card, CardContent, CardHeader, CardTitle } from "~/shared/ui/card"
+import { Progress } from "~/shared/ui/progress"
 import { Code2, Database, Wrench } from "lucide-react"
 
 const skillCategories = [
